refactor: move ServiceOrchestrator files to async_infrastructure
- Move ktrdr/managers/base.py → ktrdr/async_infrastructure/service_orchestrator.py
- Move ktrdr/managers/async_host_service.py → ktrdr/async_infrastructure/async_host_service.py
- Update all import statements across codebase (17 files updated)
- Remove generic ktrdr/managers/ directory
- Update async_infrastructure __init__.py exports

This reorganization places ServiceOrchestrator and AsyncHostService in the
more appropriate async_infrastructure directory alongside other async
components (progress, cancellation, time_estimation), improving code
organization and architectural clarity.
