+++ ktrdr/async_infrastructure/__init__.py
@@ -7,4 +7,11 @@ management across all KTRDR services.
 Components:
 - progress: Generic progress tracking with domain-specific renderers
 - cancellation: Unified cancellation system with ServiceOrchestrator integration
+- service_orchestrator: Base ServiceOrchestrator class for async service management
+- async_host_service: AsyncHostService for external service communication
 """
+
+from .async_host_service import AsyncHostService, HostServiceConfig
+from .service_orchestrator import ServiceOrchestrator
+
+__all__ = ["ServiceOrchestrator", "AsyncHostService", "HostServiceConfig"]
